backend/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bm25s==0.2.13
boto3==1.40.0
botocore==1.40.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.1
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.104.1
filetype==1.2.0
flower==2.0.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
griffe==1.9.0
grpcio==1.74.0
grpcio-health-checking==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
html2text==2025.4.15
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kombu==5.5.4
langdetect==1.0.9
llama-cloud==0.1.35
llama-cloud-services==0.6.53
llama-index==0.13.0
llama-index-cli==0.5.0
llama-index-core==0.13.0
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.0
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.0
llama-index-readers-file==0.5.0
llama-index-readers-llama-parse==0.5.0
llama-index-retrievers-bm25==0.6.0
llama-index-vector-stores-weaviate==1.4.0
llama-index-workflows==1.2.0
llama-parse==0.6.53
lxml==6.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.16
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
olefile==0.47
openai==1.98.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyStemmer==2.2.0.3
pytest==7.4.4
pytest-asyncio==0.23.2
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-oxmsg==0.0.2
python-slugify==8.0.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==5.0.1
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.27.0
striprtf==0.0.26
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.9.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.3
unstructured-client==0.41.0
urllib3==1.26.20
uvicorn==0.24.0
uvloop==0.21.0
validators==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
weaviate-client==4.16.4
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
